Changed the names of unencrypted (http) pii attacks to better suit their purpose. The names were changed to - cleartextpiiquerystring, cleartextpiihttpheader and cleartextpiihttpbody.
In GCE configuration file (mitm.conf) add data attacks parameter added the three cleartext pii attacks.

Moved the HttpDetectionMessageBodyHandler class from httppii.py to http.py.
