+++ docs/gce/mitm.conf
@@ -9,7 +9,7 @@
 #port=8080
 #attacks=selfsigned invalidhostname
 #data=httpdetection httpauthdetection
-data=piiquerystringdetection piihttpheaderdetection piihttpbodydetection
+data=cleartextpiiquerystring cleartextpiihttpheader cleartextpiihttpbody
 
 probability=0.1
 debug=True
